Move moderation opts into a context
